move version imports to inside utils
