+++ providers/src/airflow/providers/standard/sensors/time.py
@@ -21,7 +21,7 @@ import datetime
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, NoReturn
 
-from airflow.providers.standard import AIRFLOW_V_2_10_PLUS
+from airflow.providers.standard.utils.version_references import AIRFLOW_V_2_10_PLUS
 from airflow.sensors.base import BaseSensorOperator
 
 try:
